Adds the CardFactory class, which generates cards on the fly.
An associated default.png has been added to be used as a background.
Additionally, a test has been added. However, due to the difficulty of
automating the check for whether the generated card image is "acceptable,"
The test outputs 3 pngs into a tests folder under PlanningPoker for the
developer to view. The .gitignore has been updated to ignore this test
folder.

More thought will be put into the holy grail of cards, the "400 of
Diamonds"